крестики - нолики окончены. задание 2.4 номер 2
